+++ src/neuroconv/datainterfaces/__init__.py
@@ -83,6 +83,8 @@ from .ophys.caiman.caimandatainterface import CaimanSegmentationInterface
 from .ophys.cnmfe.cnmfedatainterface import CnmfeSegmentationInterface
 from .ophys.extract.extractdatainterface import ExtractSegmentationInterface
 from .ophys.hdf5.hdf5datainterface import Hdf5ImagingInterface
+from .ophys.inscopix.inscopiximagingdatainterface import InscopixImagingInterface
+from .ophys.inscopix.inscopixsegmentationdatainterface import InscopixSegmentationInterface
 from .ophys.micromanagertiff.micromanagertiffdatainterface import (
     MicroManagerTiffImagingInterface,
 )
@@ -166,6 +168,8 @@ interface_list = [
     MiniscopeImagingInterface,
     TDTFiberPhotometryInterface,
     ThorImagingInterface,
+    InscopixSegmentationInterface,
+    InscopixImagingInterface,
     # Behavior
     VideoInterface,
     ExternalVideoInterface,